+++ src/components/common/modal/index.ts
@@ -1,3 +1,2 @@
 export { default as BottomModal } from './BottomModal';
 export { default as ConfirmModal } from './ConfirmModal';
-export * from './FeedDetailModalOptions';
