+++ src/hooks/index.ts
@@ -3,5 +3,6 @@ export { useFeedDetail } from './useFeedDetail';
 export { useFeedImages } from './useFeedImages';
 export { useFeedList } from './useFeedList';
 export { useFeedWrite } from './useFeedWrite';
+export { useFeedBottomSheet } from './useFeedBottomSheet'
 export { useTimer } from './useTimer';
 export * from './modal';
