+++ src/hooks/modal/index.ts
@@ -1,4 +1,3 @@
 export * from './useAuthModal';
 export * from './useFeedCategoryModal';
-export * from './useFeedModal';
 export * from './useModal';
